requirements.txt
Django==2.1.2
requests
youtube_dl
djangorestframework